Capture draggingHeight before hiding the dragging element.
This fixes a bug where draggingHeight would be misreported if using
`display: flex` on list items, because hiding the element changes the
display to `none`, and jQuery's outerHeight() function reads the height
as if the element was `display: block` rather than `display: flex`.

At the very least this issue exists in Chrome 36; I haven't tried it in
other browsers.